more convinient beanshell chain
